Tests/LibWeb/Text/expected/css/slotted-flex-item-negative-margin.txt

wrapper width: 16
slotted offset: -12
field offset: 24
